add queries used in spec.ts files
