+++ projects/safe/src/lib/components/layout/layout.component.spec.ts
@@ -25,7 +25,7 @@ import {
   ApolloTestingModule,
   ApolloTestingController,
 } from 'apollo-angular/testing';
-import { GET_NOTIFICATIONS } from '../../graphql/queries';
+import { GET_NOTIFICATIONS } from './graphql/queries';
 import { NOTIFICATION_SUBSCRIPTION } from '../../graphql/subscriptions';
 
 describe('SafeLayoutComponent', () => {
