+++ projects/safe/src/lib/components/query-builder/query-builder.component.spec.ts
@@ -4,7 +4,7 @@ import {
   ApolloTestingModule,
   ApolloTestingController,
 } from 'apollo-angular/testing';
-import { GET_QUERY_TYPES } from '../../graphql/queries';
+import { GET_QUERY_TYPES } from './graphql/queries';
 import { SafeQueryBuilderComponent } from './query-builder.component';
 
 describe('SafeQueryBuilderComponent', () => {
