+++ projects/safe/src/lib/components/query-builder/tab-fields/tab-fields.component.spec.ts
@@ -12,7 +12,7 @@ import {
   ApolloTestingModule,
   ApolloTestingController,
 } from 'apollo-angular/testing';
-import { GET_QUERY_TYPES } from '../../../graphql/queries';
+import { GET_QUERY_TYPES } from '../graphql/queries';
 
 describe('SafeTabFieldsComponent', () => {
   let component: SafeTabFieldsComponent;
