+++ projects/safe/src/lib/components/ui/aggregation-builder/aggregation-builder.component.spec.ts
@@ -13,7 +13,7 @@ import {
   ApolloTestingModule,
   ApolloTestingController,
 } from 'apollo-angular/testing';
-import { GET_QUERY_TYPES } from '../../../graphql/queries';
+import { GET_QUERY_TYPES } from './graphql/queries';
 
 describe('SafeAggregationBuilderComponent', () => {
   let component: SafeAggregationBuilderComponent;
