+++ projects/safe/src/lib/components/ui/aggregation-builder/graphql/queries.ts
@@ -42,3 +42,123 @@ export interface GetFormsQueryResponse {
     totalCount: number;
   };
 }
+
+// === GET QUERY TYPES ===
+
+/** Graphql request for getting query types */
+export const GET_QUERY_TYPES = gql`
+  query GetQueryTypes {
+    __schema {
+      types {
+        name
+        kind
+        fields {
+          name
+          args {
+            name
+            type {
+              name
+              kind
+              inputFields {
+                name
+                type {
+                  name
+                  kind
+                }
+              }
+            }
+          }
+          type {
+            name
+            kind
+            fields {
+              name
+              args {
+                name
+                type {
+                  name
+                  kind
+                  inputFields {
+                    name
+                    type {
+                      name
+                      kind
+                    }
+                  }
+                }
+              }
+              type {
+                name
+                kind
+                ofType {
+                  name
+                  fields {
+                    name
+                    type {
+                      name
+                      kind
+                      ofType {
+                        name
+                      }
+                    }
+                  }
+                }
+              }
+            }
+            ofType {
+              name
+              fields {
+                name
+                type {
+                  name
+                  kind
+                  ofType {
+                    name
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+      queryType {
+        name
+        kind
+        fields {
+          name
+          args {
+            name
+            type {
+              name
+              kind
+              inputFields {
+                name
+                type {
+                  name
+                  kind
+                }
+              }
+            }
+          }
+          type {
+            name
+            kind
+            ofType {
+              name
+              fields {
+                name
+                type {
+                  name
+                  kind
+                  ofType {
+                    name
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+`;
