+++ projects/safe/src/lib/components/ui/core-grid/core-grid.component.spec.ts
@@ -20,7 +20,7 @@ import {
   ApolloTestingModule,
   ApolloTestingController,
 } from 'apollo-angular/testing';
-import { GET_QUERY_TYPES } from '../../../graphql/queries';
+import { GET_QUERY_TYPES } from './graphql/queries';
 import { QueryBuilderService } from '../../../services/query-builder.service';
 import {
   TranslateModule,
